+++ actions/beach-actions.ts
@@ -3,12 +3,12 @@ export { getBeaches, getBeachById } from "./beach/beach-query-actions";
 
 export {
   getNearbyBeaches,
-  calculateDistanceInMiles,
-  toRadians,
 } from "./beach/beach-location-actions";
 
 export {
   getFavoriteBeaches,
   addFavoriteBeach,
   removeFavoriteBeach,
+  reorderFavoriteBeaches,
+  getTopFavoriteBeach,
 } from "./beach/beach-favorite-actions";